playscene debreif bug

+++ src/ui.h
@@ -6,7 +6,8 @@
 
 //convenient bro
 // total (l)ength, (p)adding, (u)nit length, total (n)umber of members, (i)ndex
-#define space(l,p,u,n,i) ((p+((l-(2*p))/n)*i)+(((l-(2*p))/n)/2)-(u/2))
+//#define space(l,p,u,n,i) ((p+((l-(2.0*p))/n)*i)+(((l-(2.0*p))/n)/2.0)-(u/2.0))
+#define space(l,p,u,n,i) ((p+(((float)l-(2.0*p))/(float)n)*i)+((((float)l-(2.0*p))/(float)n)/2.0)-(u/2.0))
 
 struct In;
 class Graphics;
